Disentangle packet dissection functionally in tcpdump from the internal libnetdissect.  This greatly narrows the public interface and allows libnetdissect to be more easily sandboxed.
