+++ Makefile.in
@@ -86,6 +86,7 @@ LIBNETDISSECT_SRC=\
 	nlpid.c \
 	oui.c \
 	parsenfsfh.c \
+	print.c \
 	print-802_11.c \
 	print-802_15_4.c \
 	print-ah.c \
